Fix bug

+++ src/Enum/Combustible.php
@@ -22,10 +22,10 @@ enum Combustible: string
         $data = $reader->find($this->value);
 
         return new GES(
-            kgCO2equivalentPerLiter: (int) $data[0]['Total poste non décomposé'],
-            kgCO2fPerLiter: (int) array_values($data[0])[0],
-            kgCH4PerLiter: (int) $data[0]['CH4f'],
-            kgN2OPerLiter: (int) $data[0]['N2O']
+            kgCO2equivalentPerLiter: (float) $data[0]['Total poste non décomposé'],
+            kgCO2fPerLiter: (float) array_values($data[0])[0],
+            kgCH4PerLiter: (float) $data[0]['CH4f'],
+            kgN2OPerLiter: (float) $data[0]['N2O']
         );
     }
 }
